[YS-17440] Add support for custom keyboard type settings
Unity InputField has a special "custom" ContentType that allows
finer-grained control over the keyboard and validation settings. This
content type wasn't properly supported by the plugin, as it never
communicated the custom keyboard type to the native side. This has
now been fixed.

Android native implementation already had handling for the custom
content type, so that was not modified. However, on iOS the keyboard
type was always set based on the content type. This code has now been
modified to take the Unity-determined keyboard type into use directly.
There we can rely on the fact that Unity InputField class already
determines the right keyboard to use in all cases.